// Source/SuperBasic.Editor/Interop/Generated/JSInteropTypes.Generated.ts
// <copyright file="JSInteropTypes.Generated.ts" company="2018 Omar Tawfik">
// Copyright (c) 2018 Omar Tawfik. All rights reserved. Licensed under the MIT License. See LICENSE file in the project root for license information.
// </copyright>

/// <summary>
/// This file is auto-generated by a build task. It shouldn't be edited by hand.
/// </summary>
import { ShapesInterop } from "../JS/ShapesInterop";

export interface IShapesInterop {
    initialize(containerId: string): Promise<boolean>;
    addButton(caption: string, left: number, top: number): Promise<string>;
    addTextBox(isMultiLine: boolean, left: number, top: number): Promise<string>;
    getText(controlName: string): Promise<string>;
    hideControl(controlName: string): Promise<boolean>;
    moveControl(controlName: string, x: number, y: number): Promise<boolean>;
    removeControl(controlName: string): Promise<boolean>;
    setControlText(controlName: string, text: string): Promise<boolean>;
    setControlSize(controlName: string, width: number, height: number): Promise<boolean>;
    showControl(controlName: string): Promise<boolean>;
}

export function initializeGlobalNamespace(): void {
    (<any>global).JSIntrop = {
        Shapes: <IShapesInterop>new ShapesInterop()
    };
}